Frizzos update
The survival legendary ring (Frizzo), reapplies lacerate to the target hit with the lowest lacerate time left. This makes it viable to use carve/lacerate on single target as well.

Also carve/butchery no longer spreads a copy of itself to another target. Instead it applies a fresh instance of lacerate